+++ changelog.md
@@ -16,6 +16,10 @@
 * Fix - Sensei classes will be called from the parent LSX not child themes.
 * Dev - Allowing the sitemap loops to show items to a depth of 3.
 * Fix - Fix for 'Content wider than screen' Google Console issue.
+* Dev - Adding better breadcrumbs for woocommerce categories and taxonomies.
+* Dev - Adding simple title banner to WC pages (not product pages).
+* Dev - Updating Event Calendar Organizer styles.
+* Dev - Adding hierarchy for the sitemap.
 
 ### 2.5.1
 * Fix - Adding in a conditional to check if an interface exists before trying to include the Yoast Schema